LayoutTests/fast/js/kde/delete-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a = 1; delete a; is true
PASS delete nonexistant; is true
PASS delete NaN is false
PASS successfullyParsed is true

TEST COMPLETE
